discord_bots/voting_bot/requirements.txt
discord.py>=2.3
pyyaml
